Added initial int0 vector in assembly
The problem was that it appeared that it was taking too long to
execute the interrupt handler in c and so the int0 interrupt would
get called before int0 was called 8 times and reset the read mask.

The solution was to write the ISR for INT0 in assembly, trying to
optimize as best as possible to get the instructions. Also, the clock
was changed to be a 20Mhz crystal to get us some more time. According
to http://store.curiousinventor.com/guides/PS2/, for the PS1, the
bus frequency is between 100Khz and 500Khz. This gives about 8-24 ish
cycles before the next clock begins. In addition, the send queue was
abandoned in favor of using a single byte and mask for both send and
receive. For simplicity, it is assumed that when the mask is nonzero,
it is sending/receiving. Once the mask is entirely shifted off, the
byte has been sent/received, and so the ISR can ignore anything past
that point.