Implemented mode switch via start + select
The way it works is that when in emulator mode, PD0 is lit. When
in PSX mode, PD1 is lit. Pressing both start and select causes both
lights to be lit. After 3 seconds, if the buttons have not been
released, the mode will be switched and the other LED that was not
lit will be lit and the other one will be turned off.